Move watchlist to dashboard
