+++ client/src/app/dashboard/dashboard/page.tsx
@@ -8,6 +8,7 @@ import TradeForm from './components/TradeForm';
 import TradeHistory from './components/TradeHistory';
 import { PerformanceChart } from './components/PerformanceChart';
 import { RiskToReward } from './components/RiskToReward';
+import { Watchlist } from './components/Watchlist';
 
 export default function DashboardPage() {
     const [refreshTrigger, setRefreshTrigger] = useState(true);
@@ -53,6 +54,10 @@ export default function DashboardPage() {
                             </div>
                         </div>
 
+                    </div>
+
+                    <div className='card-container'>
+                        <Watchlist />
                     </div>
 
                     <div className='risk-reward-container card-container'>
