+++ client/src/app/styles/dashboard/watchlist.css
@@ -5,11 +5,7 @@
   flex-direction: column;
   padding: 24px;
   box-sizing: border-box;
-  background-color: #f9fafb;
-}
-
-.watchlist-header {
-  margin-bottom: 24px;
+  overflow: hidden;
 }
 
 .watchlist-header h1 {
