+++ client/src/components/NavBar.tsx
@@ -64,14 +64,6 @@ export default function NavBar() {
               <span>Trades</span>
             </Link>
             <Link
-              href="/dashboard/watchlist"
-              className={isActive('/dashboard/watchlist') ? 'active' : ''}
-              onClick={closeMobileMenu}
-            >
-              <FaEye className="nav-icon" />
-              <span>Watchlist</span>
-            </Link>
-            <Link
               href="/dashboard/settings"
               className={isActive('/dashboard/settings') ? 'active' : ''}
               onClick={closeMobileMenu}
